refactor: simplify image-only link check
fixes images from being consumed as item mentions.

cleanup: move Lexical node utility to appropriate file
