+++ lib/lexical/mdast/shared.js
@@ -1,5 +1,3 @@
-import { $isDecoratorNode, $isElementNode } from 'lexical'
-
 /**
  * checks if a node is a parent node
  * @param {Object} node - mdast node
@@ -10,10 +8,10 @@ export function isParent (node) {
 }
 
 /**
- * checks if a node is a block element
- * @param {Object} node - lexical node
- * @returns {boolean} true if node is a block element
+ * checks if a node is a link whose sole child is an image
+ * @param {Object} node - mdast node
+ * @returns {boolean} true if node is a link wrapping only an image
  */
-export function isBlockElement (node) {
-  return ($isElementNode(node) || $isDecoratorNode(node)) && !node.isInline()
+export function isImageOnlyLink (node) {
+  return node?.type === 'link' && node.children?.length === 1 && node.children[0].type === 'image'
 }
