+++ lib/lexical/mdast/transforms/links.js
@@ -1,6 +1,7 @@
 import { visit } from 'unist-util-visit'
 import { toString } from 'mdast-util-to-string'
 import { isMisleadingLink } from '@/lib/url'
+import { isImageOnlyLink } from '@/lib/lexical/mdast/shared'
 
 /**
  * a link is misleading if the text is not the same as the URL.
@@ -11,8 +12,7 @@ import { isMisleadingLink } from '@/lib/url'
 export function misleadingLinkTransform (tree) {
   visit(tree, 'link', (node, index, parent) => {
     // if the link has only an image child, unwrap it (remove the link but keep the image)
-    const hasOnlyImage = node.children?.length === 1 && node.children[0].type === 'image'
-    if (hasOnlyImage && parent && index !== undefined) {
+    if (isImageOnlyLink(node) && parent && index !== undefined) {
       parent.children[index] = node.children[0]
       return index
     }
