+++ lib/lexical/mdast/transforms/mentions.js
@@ -2,6 +2,7 @@ import { findAndReplace } from 'mdast-util-find-and-replace'
 import { visit } from 'unist-util-visit'
 import { toString } from 'mdast-util-to-string'
 import { parseInternalLinks } from '@/lib/url'
+import { isImageOnlyLink } from '@/lib/lexical/mdast/shared'
 
 // regexes from rehype-sn.js
 const userGroup = '[\\w_]+'
@@ -47,6 +48,10 @@ export function mentionTransform (tree) {
 export function itemMentionTransform (tree) {
   visit(tree, 'link', (node, index, parent) => {
     if (!node.url || !parent || index === undefined) return
+
+    // skip if link wraps only an image
+    if (isImageOnlyLink(node)) return
+
     try {
       const { itemId, commentId } = parseInternalLinks(node.url)
       if (itemId || commentId) {
