+++ lib/lexical/mdast/visitors/list.js
@@ -1,5 +1,5 @@
 import { $createListNode, $createListItemNode, $isListNode, $isListItemNode } from '@lexical/list'
-import { isBlockElement } from '@/lib/lexical/mdast/shared'
+import { $isBlockElement } from '@/lib/lexical/nodes/utils'
 
 // mdast -> lexical: list
 export const MdastListVisitor = {
@@ -54,7 +54,7 @@ const isParentCheckList = (lexicalNode) => {
 
 const createListItemNode = ({ lexicalNode, children }) => ({
   type: 'listItem',
-  spread: children.some(isBlockElement),
+  spread: children.some($isBlockElement),
   checked: isParentCheckList(lexicalNode) ? Boolean(lexicalNode.getChecked()) : null,
   children: []
 })
@@ -62,7 +62,7 @@ const createListItemNode = ({ lexicalNode, children }) => ({
 const wrapInlineContent = ({ children, listItem, actions }) => {
   let paragraph = null
   children.forEach((child) => {
-    if (isBlockElement(child)) {
+    if ($isBlockElement(child)) {
       paragraph = null
       actions.visit(child, listItem)
       return
