+++ lib/lexical/mdast/visitors/quote.js
@@ -1,6 +1,6 @@
 import { $createQuoteNode, $isQuoteNode } from '@lexical/rich-text'
 import { $isLineBreakNode } from 'lexical'
-import { isBlockElement } from '@/lib/lexical/mdast/shared'
+import { $isBlockElement } from '@/lib/lexical/nodes/utils'
 
 // mdast -> lexical
 export const MdastQuoteVisitor = {
@@ -38,7 +38,7 @@ export const LexicalQuoteVisitor = {
         continue
       }
       // block-level elements (nested quotes, lists, code blocks, etc.) are direct children of the blockquote
-      if (isBlockElement(child)) {
+      if ($isBlockElement(child)) {
         dropTrailingEmptyParagraph()
         actions.visit(child, blockquote)
         paragraph = newParagraph()
