+++ lib/lexical/nodes/utils.js
@@ -26,6 +26,15 @@ export function $replaceNodeWithLink (node, url) {
   }
 }
 
+/**
+ * checks if a Lexical node is a block element
+ * @param {Object} node - lexical node
+ * @returns {boolean} true if node is a block element
+ */
+export function $isBlockElement (node) {
+  return ($isElementNode(node) || $isDecoratorNode(node)) && !node.isInline()
+}
+
 export function $isUnwritable (node) {
   if ($isDecoratorNode(node) && !node.isInline()) return true
   if ($isElementNode(node)) {
